Defines a distributed watershed

Function to initialize an object of DistributedWatershed-class.

DistributedWatershed(spxt, lct, forestlist, soillist, bedrock, channel, 
                     validateOutlets = TRUE)

Arguments

spxt

An object of class SpatialPixelsTopography.

lct

A character vector with the land cover type of each grid cell (values should be 'wildland', 'agriculture', 'rock', 'artificial' or 'water').

forestlist

A list of objects of class 'forest' with the same number of elements as spatial pixels.

soillist

A list of objects of class 'forest' with the same number of elements as spatial pixels.

bedrock

A data frame with the same number of rows as spatial pixels and containing hydraulic bedrock properties: 'DepthToBedrock' (mm), 'Conductivity' (m/day), 'Porosity' ([0-1]).

channel

A logical vector to indicate which watershed cells belong to the channel network.

validateOutlets

A flag to remove outlet cells that are outside channel network and on the boundaries of the watershed.

Value

An object of class DistributedWatershed-class.

Author

Miquel De Cáceres Ainsa, CREAF

See also

Examples

if (FALSE) {
code = "20054998"
spt = readRDS(paste0("D:/Datasets/Hydrology/Products/Catalonia/SpatialPixelsTopography/",
                     code,".rds"))
lct = readRDS(paste0("D:/Datasets/Hydrology/Products/Catalonia/LandUse/",code,".rds"))
forestlist = readRDS(paste0("D:/Datasets/Hydrology/Products/Catalonia/Forests/",code,".rds"))
soillist = readRDS(paste0("D:/Datasets/Hydrology/Products/Catalonia/Soils/",code,".rds"))
bedrock = readRDS(paste0("D:/Datasets/Hydrology/Products/Catalonia/Bedrock/",code,".rds"))
channel = readRDS(paste0("D:/Datasets/Hydrology/Products/Catalonia/Channels/",code,".rds"))

dw = DistributedWatershed(spt, lct, forestlist, soillist, bedrock, channel)
}