Class "SpatialPixelsLandscape"

An S4 class that represents a landscape configuration over an (incomplete) grid of coordinates.

Objects from the Class

Objects can be created by calls of the form new("SpatialPixelsLandscape", ...).

Slots

data:

Object of class "data.frame" containing the elevation (in m), slope (in degrees) and aspect (in degrees from North) of every cell.

coords:

A matrix of pixel coordinates.

coords.nrs:

Inherited from SpatialPointsDataFrame but not used.

grid:

Object of class GridTopology.

grid.index:

Index of points in full grid.

bbox:

Object of class "matrix" with the boundary box.

proj4string:

Object of class "CRS" with the projection string.

lct:

A character vector with the land cover type of each grid cell (values should be 'wildland', 'agriculture', 'rock', 'artificial' or 'water').

forestlist:

Object of class "list" containing a set of forest objects.

soillist:

Object of class "list" containing a set of soil objects.

xlist:

Object of class "list" containing a set of spwb or growth model input objects.

Extends

Class "SpatialPixelsTopography", directly. Class "SpatialPixelsDataFrame", distance 2. Class "SpatialPixels", distance 3. Class "SpatialPointsDataFrame", distance 3. Class "SpatialPoints", distance 4. Class "Spatial", distance 5.

Methods

spatialForestSummary

signature(object = "SpatialPixelsLandscape"): calculates a summary function for all forest stands and returns an object of class SpatialPixelsDataFrame-class.

spatialSoilSummary

signature(object = "SpatialPixelsLandscape"): calculates a summary function for the soil of all stands and returns an object of class SpatialPixelsDataFrame-class.

getLandscapeLayer

signature(object = "SpatialPixelsLandscape"): allows retrieving a variable as SpatialPixelsDataFrame.

getLandscapeVariable

signature(object = "SpatialPixelsLandscape"): allows retrieving a variable as vector.

spplot

signature(object = "SpatialPixelsLandscape"): allows plotting maps of the landscape state.

Author

Miquel De Cáceres Ainsa, Centre Tecnologic Forestal de Catalunya

See also

Examples

#Structure of the S4 object
showClass("SpatialPixelsLandscape")
#> Class "SpatialPixelsLandscape" [package "medfateland"]
#> 
#> Slots:
#>                                                                        
#> Name:           lct   forestlist     soillist        xlist         data
#> Class:    character         list         list         list   data.frame
#>                                                                        
#> Name:    coords.nrs         grid   grid.index       coords         bbox
#> Class:      numeric GridTopology      integer       matrix       matrix
#>                    
#> Name:   proj4string
#> Class:          CRS
#> 
#> Extends: 
#> Class "SpatialPixelsTopography", directly
#> Class "SpatialPixelsDataFrame", by class "SpatialPixelsTopography", distance 2
#> Class "SpatialPixels", by class "SpatialPixelsTopography", distance 3
#> Class "SpatialPointsDataFrame", by class "SpatialPixelsTopography", distance 3
#> Class "SpatialPoints", by class "SpatialPixelsTopography", distance 4
#> Class "Spatial", by class "SpatialPixelsTopography", distance 5
#> Class "SpatialVector", by class "SpatialPixelsTopography", distance 5
#> 
#> Known Subclasses: "DistributedWatershed"