Adds SoilGrids soil description

Fills column 'soil' with information drawn from SoilGrids 2.0

add_soilgrids(
  x,
  soilgrids_path = NULL,
  widths = NULL,
  replace_existing = TRUE,
  verbose = TRUE
)

Arguments

x

An object of class sf with a valid CRS definition. If it contains a column called 'land_cover_type', soils will be retrieved for "agriculture" and "wildland" cover types. Otherwise, soils are retrieved for all locations.

soilgrids_path

Path to SoilGrids rasters (see details). If missing, the SoilGrids REST API (https://rest.isric.org) will be queried.

widths

A numeric vector indicating the desired layer widths, in mm. If NULL the default soil grids layer definition is returned.

replace_existing

A logical flag to force the replacement of existing soil data, when already present

verbose

A logical flag to include a progress bar while processing the output of the query to the SoilGrids REST API.

Value

A modified object of class sf with column 'soil'.

Details

If soilgrids_path = NULL the function connects with the SoilGrids REST API (https://rest.isric.org) to retrieve the soil physical and chemical characteristics for a site (Hengl et al. 2007), selected by its coordinates. Also, in case the depths are not the default ones in the SoilGrids API, the function uses averages the values of soil grid layers depending on the overlap between soil layer definitions.

If soilgrids_path != NULL the function will read SoilGrid rasters from the file disk. Folders need to be defined for each variable ("sand", "clay", "soc", "bdod", "cfvo" and "nitrogen"). File paths from soilgrids_path should be named:

var/var_layer_mean.tif

where var is one of the above and layer is "0-5cm", "5-15cm", "15-30cm", "30-60cm", "60-100cm" or "100-200cm"

References

Hengl T, Mendes de Jesus J, Heuvelink GBM, Ruiperez Gonzalez M, Kilibarda M, Blagotić A, et al. (2017) SoilGrids250m: Global gridded soil information based on machine learning. PLoS ONE 12(2): e0169748. doi:10.1371/journal.pone.0169748.

See also

Author

Víctor Granda, EMF-CREAF

Miquel De Cáceres Ainsa, EMF-CREAF

Examples

 if (FALSE) {
  # See vignette 'Landscape inputs'
 }