Seed production, dispersal and seed bank dynamics

Simulates seed bank mortality, seed production and dispersal among stands

dispersal(
  sf,
  SpParams,
  local_control = medfate::defaultControl(),
  distance_step = 25,
  maximum_dispersal_distance = 3000,
  min_percent = 1,
  stochastic_resampling = FALSE,
  progress = TRUE
)

Arguments

sf

An object of class sf using a UTM projection (to measure distances in m) and with the following columns:

geometry: Spatial geometry.

forest: Objects of class forest.

SpParams

A data frame with species parameters (see SpParamsMED).

local_control

A list of control parameters (see defaultControl)

distance_step

Distance step in meters.

maximum_dispersal_distance

Maximum dispersal distance in meters.

min_percent

A minimum percent of seed bank to retain entry in seedBank element of forest.

stochastic_resampling

A flag to indicate that stochastic resampling of stands is performed.

progress

Boolean flag to display progress information.

Value

A list with forest objects (for wildland cover type) containing a modified seed bank

Details

The input 'sf' object has to be in a Universal Transverse Mercator (UTM) coordinate system (or any other projection using meters as length unit) for appropriate function behavior.

Dispersal kernel follows Clark et al. (1999) and potential seed donors (neighbors) are defined up to a given grid distance order. A maximum value of 100% seed bank refill is attained for species with seed production in all seed donors and the local cell.

References

Clark et al. (1999). Seed dispersal near and far: Patterns across temperate and tropical forests. Ecology 80(5):1475-1494

See also

Author

Miquel De Cáceres Ainsa, CREAF.

Roberto Molowny-Horas, CREAF.

Examples

if (FALSE) {
data(example_watershed)
data(SpParamsMED)

# Transform to UTM31
example_watershed_utm31 <- sf::st_transform(example_watershed, crs = 32631)

# Estimate seed production and dispersal over the watershed
seedbank_list <- dispersal(example_watershed_utm31, SpParamsMED)

seedbank_list[[1]]

# Transform to UTM31
example_ifn_utm31 <- sf::st_transform(example_ifn, crs = 32631)

# Estimate seed production and dispersal over the set of forest inventory plots
seedbank_list <- dispersal(example_ifn_utm31, SpParamsMED)

seedbank_list[[1]]
}