Fire regime instance

Applies a fire regime object over a set of landscape units to determine a fire realization

fire_regime_instance(sf, fire_regime)

Arguments

sf

An object of class sf with the following columns:

geometry: Spatial geometry.

id: Stand identifiers.

represented_area: Area represented by each stand (in hectares).

ignition_weights: Relative weights to determine stands to be burned (optional).

fire_regime

A list of parameters defining the fire regime (see create_fire_regime).

Value

An integer matrix specifying the day of the year of burning of each landscape unit for every year in the fire regime definition. Values are interpreted as follows:

NA - No wildfire this year

0 - Wildfire will occur the driest day (i.e. the one with largest vapor pressure deficit).

1...366 - Day of the year when wildfire will occur

Details

The function randomly determines the landscape units that will burn every year, depending on the specifications of the fire regime object.

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

# Load example data
data("example_ifn")

# Assume that each stand represents 1km2 = 100 ha
example_ifn$represented_area <- 100

# Define fire regime characteristics
freg <- create_fire_regime(c("2002" = 200, "2003" = 500)) 

# Create a fire regime instance
fire_regime_instance(example_ifn, freg)
#>            2002 2003
#> 081015_A1    NA   NA
#> 081016_A1     0   NA
#> 081018_A1    NA   NA
#> 081019_A1    NA   NA
#> 081020_A1    NA   NA
#> 081021_A1    NA   NA
#> 081022_A1    NA   NA
#> 081023_A1    NA   NA
#> 081024_A1    NA   NA
#> 081025_A1    NA   NA
#> 081026_A1    NA   NA
#> 081027_A1    NA    0
#> 081028_A1    NA   NA
#> 081029_A1    NA   NA
#> 081030_A1    NA   NA
#> 081031_A1    NA   NA
#> 081032_A1    NA   NA
#> 081033_A1    NA   NA
#> 081034_A1    NA   NA
#> 081035_A1    NA    0
#> 081036_A1    NA   NA
#> 081037_A1    NA   NA
#> 081038_A1    NA   NA
#> 081039_A1    NA   NA
#> 081040_A1    NA   NA
#> 081041_A1    NA    0
#> 081043_A1    NA   NA
#> 081044_A1    NA   NA
#> 081045_A1    NA   NA
#> 081046_A1    NA   NA
#> 081047_A1    NA   NA
#> 081048_A1    NA   NA
#> 081049_A1    NA   NA
#> 081050_A1    NA   NA
#> 081051_A1    NA   NA
#> 081052_A1    NA   NA
#> 081053_A1    NA   NA
#> 081054_A1    NA   NA
#> 081055_A1    NA   NA
#> 081056_A1    NA   NA
#> 081057_A1    NA   NA
#> 081058_A1    NA   NA
#> 081059_A1    NA   NA
#> 081060_A1    NA   NA
#> 081061_A1    NA   NA
#> 081062_A1    NA    0
#> 081063_A1    NA   NA
#> 081064_A1    NA   NA
#> 081065_A1    NA   NA
#> 081066_A4    NA   NA
#> 081067_A1    NA   NA
#> 081068_A1    NA   NA
#> 081069_A1    NA   NA
#> 081070_A1    NA   NA
#> 081071_A4    NA   NA
#> 081073_A1    NA   NA
#> 081075_A4    NA   NA
#> 081076_A1    NA    0
#> 081077_A1    NA   NA
#> 081078_A4C   NA   NA
#> 081079_A1    NA   NA
#> 081080_A1    NA   NA
#> 081081_A1    NA   NA
#> 081082_A1    NA   NA
#> 081083_A1    NA   NA
#> 081084_A1    NA   NA
#> 081085_A1     0   NA
#> 081086_A4    NA   NA
#> 081087_A4    NA   NA
#> 081088_A1    NA   NA
#> 081089_A3C   NA   NA
#> 081089_A3E   NA   NA
#> 081090_A1    NA   NA
#> 081091_A1    NA   NA
#> 081092_A1    NA   NA
#> 081093_A1    NA   NA
#> 081094_A1    NA   NA
#> 081095_A1    NA   NA
#> 081096_A1    NA   NA
#> 081097_A1    NA   NA
#> 081098_A1    NA   NA
#> 081099_A1    NA   NA
#> 081100_A1    NA   NA
#> 081101_A1    NA   NA
#> 081102_A1    NA   NA
#> 081103_A4    NA   NA
#> 081104_A1    NA   NA
#> 081105_A1    NA   NA
#> 081106_A3C   NA   NA
#> 081106_A3E   NA   NA
#> 081107_A4    NA   NA
#> 081108_A1    NA   NA
#> 081109_A1    NA   NA
#> 081110_A1    NA   NA
#> 081111_A1    NA   NA
#> 081112_A1    NA   NA
#> 081113_A4    NA   NA
#> 081114_A1    NA   NA
#> 081115_A1    NA   NA
#> 081116_A4    NA   NA