Scenario of forest dynamics

Evaluates forest dynamics over a landscape including climate and management scenarios

fordyn_scenario(
  y,
  SpParams,
  meteo,
  volumeFunction,
  managementScenario,
  localControl = defaultControl(),
  dates = NULL,
  CO2ByYear = numeric(0),
  parallelize = FALSE,
  numCores = detectCores() - 1,
  chunk.size = NULL,
  progress = TRUE
)

Arguments

y

An object of class sf with landscape information. Management units should be defined in this object.

SpParams

A data frame with species parameters (see SpParamsMED).

meteo

Meteorology data (see fordyn_spatial).

volumeFunction

A function accepting a forest object as input and returning the wood volume corresponding to each tree cohort.

managementScenario

A list defining the management scenario (see create_management_scenario)

localControl

A list of local model control parameters (see defaultControl).

dates

A Date object with the days of the period to be simulated. If NULL, then the whole period of meteo is used.

CO2ByYear

A named numeric vector with years as names and atmospheric CO2 concentration (in ppm) as values. Used to specify annual changes in CO2 concentration along the simulation (as an alternative to specifying daily values in meteo).

parallelize

Boolean flag to try parallelization (will use all clusters minus one).

numCores

Integer with the number of cores to be used for parallel computation.

chunk.size

Integer indicating the size of chunks to be sent to different processes (by default, the number of spatial elements divided by the number of cores).

progress

Boolean flag to display progress information for simulations.