Forest and soil summaries over space

Functions to calculates a summary function for the forest or soil of all spatial elements in an object of class sf containing landscape information.

landscape_summary(object, name, summaryFunction, ..., unlist = FALSE)

Arguments

object

An object of class sf.

name

A string of the element to summarize: "forest", "soil" or "state".

summaryFunction

A function that accepts objects of class forest, soil or model input objects, respectively.

...

Additional arguments to the summary function.

unlist

Logical flag to try converting the summaries into different columns

Value

An object of class sf containing the calculated statistics. If unlist = FALSE column 'summary' is a list with summaries for each element. If unlist = TRUE different columns are returned instead, one per variable given in the summary function.

See also

Author

Miquel De Cáceres Ainsa, CREAF.

Examples

# Load plot data and species parameters from medfate
data(examplepointslandscape)

# Load default medfate parameters
data("SpParamsMED")
 
# Transform example to 'sf' 
y = sp_to_sf(examplepointslandscape)

# Apply forest summary function
landscape_summary(y, "forest", summary.forest, SpParamsMED)
#> Simple feature collection with 100 features and 1 field
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 402000 ymin: 4640000 xmax: 429000 ymax: 4646000
#> Projected CRS: ETRS89 / UTM zone 31N
#> # A tibble: 100 × 2
#>            geometry summary        
#>         <POINT [m]> <list>         
#>  1 (411000 4646000) <smmry.fr [13]>
#>  2 (412000 4646000) <smmry.fr [13]>
#>  3 (413000 4646000) <smmry.fr [13]>
#>  4 (415000 4646000) <smmry.fr [13]>
#>  5 (416000 4646000) <smmry.fr [13]>
#>  6 (423000 4646000) <smmry.fr [13]>
#>  7 (426000 4646000) <smmry.fr [13]>
#>  8 (428000 4646000) <smmry.fr [13]>
#>  9 (402000 4645000) <smmry.fr [13]>
#> 10 (404000 4645000) <smmry.fr [13]>
#> # … with 90 more rows