Medfate simulations for spatially-distributed forest stands

Functions that allow calling local models spwb, growth or fordyn, for a set of forest stands distributed in specific locations. No spatial processes are simulated.

spwbpoints(y, SpParams, meteo, control = defaultControl(),
       dates = NULL, keepResults = TRUE, summaryFunction = NULL, args=NULL, 
       parallelize = FALSE, progress = TRUE)
spwbpixels(y, SpParams, meteo, control = defaultControl(),
       dates = NULL, keepResults = TRUE, summaryFunction = NULL, args=NULL, 
       parallelize = FALSE, progress = TRUE)
spwbgrid(y, SpParams, meteo, control = defaultControl(),
       dates = NULL, keepResults = TRUE, summaryFunction = NULL, args=NULL, 
       parallelize = FALSE, progress = TRUE)
growthpoints(y, SpParams, meteo, control = defaultControl(),
       dates = NULL, keepResults = TRUE, summaryFunction = NULL, args=NULL, 
       parallelize = FALSE, progress = TRUE)
growthpixels(y, SpParams, meteo, control = defaultControl(),
       dates = NULL, keepResults = TRUE, summaryFunction = NULL, args=NULL, 
       parallelize = FALSE, progress = TRUE)
growthgrid(y, SpParams, meteo, control = defaultControl(),
       dates = NULL, keepResults = TRUE, summaryFunction = NULL, args=NULL, 
       parallelize = FALSE, progress = TRUE)
fordynpoints(y, SpParams, meteo, control = defaultControl(),
       dates = NULL, keepResults = TRUE, summaryFunction = NULL, args=NULL, 
       parallelize = FALSE, progress = TRUE)
fordynpixels(y, SpParams, meteo, control = defaultControl(),
       dates = NULL, keepResults = TRUE, summaryFunction = NULL, args=NULL, 
       parallelize = FALSE, progress = TRUE)
fordyngrid(y, SpParams, meteo, control = defaultControl(),
       dates = NULL, keepResults = TRUE, summaryFunction = NULL, args=NULL, 
       parallelize = FALSE, progress = TRUE)

Arguments

y

An object of class SpatialPointsLandscape-class, SpatialPixelsLandscape-class or SpatialGridLandscape-class, depending on the function called.

SpParams

A data frame with species parameters (see SpParamsMED).

meteo

Meteorology data (see details).

control

A list of model control parameters (see defaultControl).

dates

A Date object with the days of the period to be simulated. If NULL, then the whole period of meteo is used.

keepResults

Boolean flag to indicate that point/cell simulation results are to be returned (set to FALSE and use summary functions for large data sets).

summaryFunction

An appropriate function to calculate summaries (e.g., summary.spwb).

args

List with additional arguments for the summary function.

paralellize

Boolean flag to try parallelization (will use all clusters minus one).

progress

Boolean flag to display progress bar for simulations.

Details

Simulation functions accept different formats for meteorological input (parameter meteo). The user may supply three kinds of weather sources:

An object of SpatialPointsMeteorology-class, SpatialGridMeteorology-class or SpatialPixelsMeteorology-class, with the same spatial geometry as y.

An object of MeteorologyInterpolationData-class.

A data frame with meteorological data common for all points or grid cells.

In the case of (1), all the spatio-temporal variation of weather is already supplied by the user. In the case of (2), interpolation of weather is done for each spatial location and the period indicated in meteo or, if supplied, dates. Finally, in the case of (3) spatial variation of weather is not considered.

Value

A list of class of the same name as the function called, containing spatial information of the input data and two child lists:

xlist: A list of model input objects for each simulated stand.

resultlist: A list of model output for each simulated stand.

summarylist: A list of model output summaries for each simulated stand.

Author

Miquel De Cáceres Ainsa, CREAF

See also

Examples

if (FALSE) {
#Load example watershed (inherits from SpatialPixelsLandscape)
data("examplepointslandscape")

#Load example meteo data frame from package meteoland
data("examplemeteo")

#Load default medfate parameters
data("SpParamsMED")

dates = seq(as.Date("2001-03-01"), as.Date("2001-03-15"), by="day")

res = spwbpoints(examplepointslandscape, SpParamsMED, examplemeteo, dates = dates, 
                 summaryFunction = summary.spwb)
}