One-day simulation for spatially-distributed forest stands

Functions that allow calling local models spwb_day or growth_day, for a set of forest stands distributed in specific locations and a given date. No spatial processes are simulated.

spwbpoints_day(y, meteo, date, SpParams, localControl = defaultControl(),
       parallelize = FALSE, numCores = detectCores()-1, chunk.size = NULL, 
       progress = TRUE)
spwbpixels_day(y, meteo, date, SpParams, localControl = defaultControl(),
       parallelize = FALSE, numCores = detectCores()-1, chunk.size = NULL, 
       progress = TRUE)
spwbgrid_day(y, meteo, date, SpParams, localControl = defaultControl(),
       parallelize = FALSE, numCores = detectCores()-1, chunk.size = NULL, 
       progress = TRUE)
growthpoints_day(y, meteo, date, SpParams, localControl = defaultControl(),
       parallelize = FALSE, numCores = detectCores()-1, chunk.size = NULL, 
       progress = TRUE)
growthpixels_day(y, meteo, date, SpParams, localControl = defaultControl(),
       parallelize = FALSE, numCores = detectCores()-1, chunk.size = NULL, 
       progress = TRUE)
growthgrid_day(y, meteo, date, SpParams, localControl = defaultControl(), 
       parallelize = FALSE, numCores = detectCores()-1, chunk.size = NULL, 
       progress = TRUE)

Arguments

y

An object of class SpatialPointsLandscape-class, SpatialPixelsLandscape-class or SpatialGridLandscape-class, depending on the function called.

meteo

Meteorology data (see spwbpoints).

date

A string with the date to be simulated.

SpParams

A data frame with species parameters (see SpParamsMED).

localControl

A list of local model control parameters (see defaultControl).

parallelize

Boolean flag to try parallelization (will use all clusters minus one).

numCores

Integer with the number of cores to be used for parallel computation.

chunk.size

Integer indicating the size of chuncks to be sent to different processes (by default, the number of spatial elements divided by the number of cores).

progress

Boolean flag to display progress information for simulations.

Details

Simulation functions accept different formats for meteorological input (parameter meteo) as described in spwbpoints

Value

A list of class of the same name as the function called containing three elements:

sp: An object with spatial information (of SpatialPoints-class, SpatialPixels-class or SpatialGrid-class).

xlist: A list of model input objects for each simulated stand, to be used in subsequent simulations.

resultlist: A list of model output for each simulated stand.

Author

Miquel De Cáceres Ainsa, CREAF

See also

Examples

if (FALSE) {
#Load example watershed (inherits from SpatialPixelsLandscape)
data("examplepointslandscape")

#Load example meteo data frame from package meteoland
data("examplemeteo")

#Load default medfate parameters
data("SpParamsMED")

#Perform simulation
date = "2001-03-01"
res = spwbpoints_day(examplepointslandscape, examplemeteo, date, SpParamsMED)
}