Watershed simulations

Functions to perform simulations on a watershed described by a set of connected grid cells

spwb_land(
  sf,
  SpParams,
  meteo = NULL,
  dates = NULL,
  CO2ByYear = numeric(0),
  summary_frequency = "years",
  local_control = medfate::defaultControl(),
  correction_factors = default_watershed_correction_factors(),
  progress = TRUE
)

growth_land(
  sf,
  SpParams,
  meteo = NULL,
  dates = NULL,
  CO2ByYear = numeric(0),
  summary_frequency = "years",
  local_control = medfate::defaultControl(),
  correction_factors = default_watershed_correction_factors(),
  progress = TRUE
)

fordyn_land(
  sf,
  SpParams,
  meteo = NULL,
  dates = NULL,
  CO2ByYear = numeric(0),
  local_control = medfate::defaultControl(),
  correction_factors = default_watershed_correction_factors(),
  management_function = NULL,
  progress = TRUE
)

Arguments

sf

An object of class sf with the following columns:

geometry: Spatial geometry.

id: Cell ids (normally row number).

elevation: Elevation above sea level (in m).

slope: Slope (in degrees).

aspect: Aspect (in degrees).

land_cover_type: Land cover type of each grid cell (values should be 'wildland', 'agriculture', 'rock', 'artificial' or 'water').

forest: Objects of class forest.

soil: Objects of class soil or data frames of physical properties.

state: Objects of class spwbInput or growthInput (optional).

meteo: Data frames with weather data (required if parameter meteo = NULL).

crop_factor: Crop evapo-transpiration factor. Only required for 'agriculture' land cover type.

waterOrder: Integer vector indicating cell processing order.

waterQ: A list of water discharge values to neighbors.

queenNeigh: A list of integers identifying the (up to 8) queen neighbors, for each cell.

channel: A logical vector indicating whether each cell belongs to the channel network.

depth_to_bedrock: Depth to bedrock (mm).

bedrock_conductivity: Bedrock (saturated) conductivity (in m·day-1).

bedrock_porosity: Bedrock porosity.

snowpack: A numeric vector with the snow water equivalent content of the snowpack in each cell.

aquifer: A numeric vector with the water content of the aquifer in each cell.

represented_area: Area represented by each cell (in m2).

management_arguments: Lists with management arguments (optional, relevant for fordyn_land only).

SpParams

A data frame with species parameters (see SpParamsMED).

meteo

Input meteorological data (see spwb_spatial).

dates

A Date object describing the days of the period to be modeled.

CO2ByYear

A named numeric vector with years as names and atmospheric CO2 concentration (in ppm) as values. Used to specify annual changes in CO2 concentration along the simulation (as an alternative to specifying daily values in meteo).

summary_frequency

Frequency in which summary layers will be produced (e.g. "years", "months", ...) (see cut.Date). In fordyn_land summaries are always produced at monthly resolution.

local_control

A list of control parameters (see defaultControl) for function spwb_day or growth_day.

correction_factors

A list of watershed correction factors for hydraulic parameters.

progress

Boolean flag to display progress information for simulations.

management_function

A function that implements forest management actions (see fordyn). of such lists, one per spatial unit.

Value

Functions spwb_land, growth_land and fordyn_land return a list of class of the same name as the function with the following elements:

sf: An object of class sf, similar to the output of spwb_spatial, with the following columns:

geometry: Spatial geometry.

id: Cell id, taken from the input.

state: A list of model input objects for each simulated stand.

aquifer: A numeric vector with the water volume in the aquifer of each cell.

snowpack: A numeric vector with the snowpack water equivalent volume of each cell.

summary: A list of cell summaries, containing the following variables:

Rain: Rainfall (in mm).

Snow: Snowfall (in mm).

Snowmelt: Snow melt (in mm).

Interception: Rainfall interception (in mm).

NetRain: Net rain, i.e. throughfall, (in mm).

Runon: The amount of water imported from other cells via surface runoff (in mm).

Runoff: The amount of water exported via surface runoff (in mm).

Infiltration: The amount of water infiltrating into the soil (in mm).

DeepDrainage: The amount of water draining from soil to the aquifer via deep drainage (in mm).

SaturationExcess: The amount of water that reaches the soil surface because of soil saturation (in mm).

AquiferDischarge: The amount of water that reaches deepest soil layer from a saturated aquifer (in mm).

SubsurfaceInput: The amount of water that reaches the soil from adjacent cells via subsurface flow (in mm).

SubsurfaceOutput: The amount of water that leaves the soil towards adjacent cells via subsurface flow (in mm).

GroundwaterInput: The amount of water that reaches the aquifer from adjacent cells via groundwater flow (in mm).

GroundwaterOutput: The amount of water that leaves the aquifer towards adjacent cells via groundwater flow (in mm).

SoilEvaporation: Bare soil evaporation (in mm).

Transpiration: Plant transpiration (in mm).

SWE: Snow water equivalent (in mm) of the snowpack.

Psi1: Soil water potential of the topmost layer (in MPa).

SoilVol: Soil water volume integrated across vertical layers (in mm).

WTD: Water table depth (in mm from surface).

In function fordyn_land the sf object contains additional columns:

forest: A list of forest objects for each simulated stand (only in fordynspatial), to be used in subsequent simulations (see update_landscape).

management_arguments: A list of management arguments for each simulated stand (only in fordynspatial), to be used in subsequent simulations (see update_landscape).

tree_table: A list of data frames for each simulated stand, containing the living trees at each time step.

shrub_table: A list of data frames for each simulated stand, containing the living shrub at each time step.

dead_tree_table: A list of data frames for each simulated stand, containing the dead trees at each time step.

dead_shrub_table: A list of data frames for each simulated stand, containing the dead shrub at each time step.

cut_tree_table: A list of data frames for each simulated stand, containing the cut trees at each time step.

cut_shrub_table: A list of data frames for each simulated stand, containing the cut shrub at each time step.

watershed_balance: A data frame with as many rows as summary points and where columns are components of the water balance at the watershed level (i.e., rain, snow, interception, infiltration, soil evaporation, plant transpiration, ...).

watershed_soil_balance: A data frame with as many rows as summary points and where columns are components of the water balance at the watershed level restricted to those cells with a soil definition.

daily_runoff: A matrix with daily runoff (in m3/day) at each of the outlet cells of the landscape.

Details

Function spwb_land implements a distributed hydrological model that simulates daily local water balance, from spwb_day, on grid cells of a watershed while accounting for overland runoff, subsurface flow and groundwater flow between cells.

Function growth_land is similar to spwb_land, but includes daily local carbon balance, growth and mortality processes in grid cells, provided by growth_day.

Function fordyn_land extends the previous two functions with the simulation of management, recruitment and resprouting.

See also

Author

Miquel De Cáceres Ainsa, CREAF.

Examples

if (FALSE) {
# Load example watershed data
data("example_watershed")

# Set crop factor 
example_watershed$crop_factor <- NA
example_watershed$crop_factor[example_watershed$land_cover_type=="agriculture"] <- 0.75

# Load example meteo data frame from package meteoland
data("examplemeteo")
  
# Load default medfate parameters
data("SpParamsMED")
  
# Set simulation period
dates <- seq(as.Date("2001-01-01"), as.Date("2001-03-31"), by="day")

# Launch simulations
res <- spwb_land(example_watershed, SpParamsMED, examplemeteo, 
                 dates = dates, summary_frequency = "month")
}