Watershed simulations

Function spwbland implements a distributed hydrological model that simulates daily local water balance, from spwb_day, on grid cells of a watershed while accounting for overland runoff, subsurface flow and groundwater flow between cells. Function growthland is similar, but includes daily local carbon balance and growth processes in grid cells, provided by growth_day.

spwbland(y, SpParams, meteo, dates,
     summaryFreq = "years",
     localControl = medfate::defaultControl(),
     correctionFactors = defaultWatershedCorrectionFactors())
growthland(y, SpParams, meteo, dates,
     summaryFreq = "years",
     localControl = medfate::defaultControl(),
     correctionFactors = defaultWatershedCorrectionFactors())

Arguments

y

An object of class DistributedWatershed-class.

SpParams

A data frame with species parameters (see SpParamsMED).

meteo

Input meteorological data (see section details).

dates

A Date object describing the days of the period to be modeled.

summaryFreq

Frequency in which summary layers will be produced (e.g. "years", "months", ...) (see cut.Date).

localControl

A list of control parameters (see defaultControl) for function spwb_day or growth_day.

correctionFactors

A list of watershed correction factors for hydraulic parameters.

Details

Functions spwbland and growthland require daily meteorological data over grid cells. The user may supply four different inputs:

An object of SpatialPixelsMeteorology-class.

An object of MeteorologyInterpolationData-class.

A data frame with information regarding where to read meteorological data.

A data frame with meteorological data common for all cells of the grid.

In the case of (1), all the spatio-temporal variation of weather is already supplied by the user. In the case of (2), interpolation of weather is done over each grid cell every simulated day. In the case of (3) weather maps are read for each day. Finally, in the case of (4) spatial variation of weather is not considered.

Value

Function spwbland list of class 'spwbland' with the following elements:

coords: A matrix of pixel coordinates.

coords.nrs: Inherited from SpatialPointsDataFrame but not used.

grid: The GridTopology object corresponding to the simulated area.

grid.index: Index of points in full grid.

bbox: Object of class "matrix" with the boundary box.

WatershedBalance: A data frame with as many rows as summary points and where columns are components of the water balance at the watershed level (i.e., rain, snow, interception, infiltration, soil evaporation, plant transpiration, ...).

WatershedSoilBalance: A data frame with as many rows as summary points and where columns are components of the water balance at the watershed level restricted to those cells with a soil definition.

CellBalance: A list of matrices (each with as many rows as cells and as many columns as summary points) representing the water balance for each cell:

Rain: Rainfall (in mm).

Snow: Snowfall (in mm).

Snowmelt: Snow melt (in mm).

Interception: Rainfall interception (in mm).

NetRain: Net rain, i.e. throughfall, (in mm).

Runon: The amount of water imported from other cells via surface runoff (in mm).

Runoff: The amount of water exported via surface runoff (in mm).

Infiltration: The amount of water infiltrating into the soil (in mm).

DeepDrainage: The amount of water draining from soil to the aquifer via deep drainage (in mm).

SaturationExcess: The amount of water that reaches the soil surface because of soil saturation (in mm).

AquiferDischarge: The amount of water that reaches deepest soil layer from a saturated aquifer (in mm).

SubsurfaceInput: The amount of water that reaches the soil from adjacent cells via subsurface flow (in mm).

SubsurfaceOutput: The amount of water that leaves the soil towards adjacent cells via subsurface flow (in mm).

GroundwaterInput: The amount of water that reaches the aquifer from adjacent cells via groundwater flow (in mm).

GroundwaterOutput: The amount of water that leaves the aquifer towards adjacent cells via groundwater flow (in mm).

SoilEvaporation: Bare soil evaporation (in mm).

Transpiration: Plant transpiration (in mm).

CellState: A list of matrices (each with as many rows as cells and as many columns as summary points) representing the average soil state for each cell:

SWE: Snow water equivalent (in mm) of the snowpack.

Psi1: Soil water potential of the topmost layer (in MPa).

SoilVol: Soil water volume integrated across vertical layers (in mm).

WTD: Water table depth (in mm from surface).

DailyRunoff: A matrix with daily runoff (in m3/day) at each of the outlet cells of the landscape.

Author

Miquel De Cáceres Ainsa, CREAF.

See also

Examples

if (FALSE) {
data("examplewatershed")
data("examplemeteo")
data("SpParamsMED")

dates = seq(as.Date("2001-01-01"), as.Date("2001-03-31"), by="day")

res = spwbland(examplewatershed, SpParamsMED, examplemeteo, dates = dates, summaryFreq = "month")
}