Add topography and land cover

Initializes topography and land cover type for a set of target locations

Usage

add_topography(x, dem, progress = TRUE)

add_land_cover(
  x,
  land_cover_map,
  wildland = NULL,
  agriculture = NULL,
  rock = NULL,
  artificial = NULL,
  water = NULL,
  progress = TRUE
)

Arguments

x

An object of class sf

dem

A digital elevation model (class SpatRaster) with meters as units

progress

A logical flag to print console output

land_cover_map

An object of class SpatRaster of land cover type. If missing, all locations are considered 'wildland'.

wildland, agriculture, rock, artificial, water

Strings indicating the mapping from the legend of land_cover_map.

Value

Function add_topography() returns a modified object of class sf with columns:

id: Numeric location identifiers (if not existing).

elevation: Elevation above sea level (in m).

slope: Slope (in degrees).

aspect: Aspect (in degrees).

land_cover_type: Land cover type.

Function add_land_cover() returns a modified object of class sf with new column:

id: Numeric location identifiers (if not existing).

land_cover_type: Land cover type.

Details

The user should manually define the mapping of land cover classes in land_cover_map to the land cover types used in medfateland.

See also

Examples

# See package vignettes 'Preparing inputs'