Default control parameters for watershed processes

Defines default control parameters for watershed processes

default_watershed_control(watershed_model = "tetis")

Arguments

watershed_model

Hydrological model for watershed processes. Only "tetis" or "serghei" are accepted.

Value

A list with the following items:

watershed_model: A string with the watershed model.

weather_aggregation_factor [= 1]: An integer specifying the spatial aggregation for interpolated weather.

tetis_parameters: A list of TETIS parameters with the following elements:

R_localflow [= 1.0]: Correction factor for soil hydraulic saturated conductivity (local vertical flows).

R_interflow [= 1.0]: Correction factor for soil hydraulic saturated conductivity (subsurface flow between grid cells).

R_baseflow [= 10.0]: Correction factor for bedrock hydraulic conductivity (groundwaterflow between grid cells).

serghei_parameters: A list of SERGHEI parameters with the following elements:

input_dir [= ""]: Path to SERGHEI input files.

output_dir [= ""]: Path to SERGHEI output files.

force_equal_layer_widths [= FALSE]: A boolean flag to force equal layer widths (taken from the first soil element) in all soils.

dispersal_parameters: A list of dispersal parameters (only for fordyn_land) with the following elements:

distance_step [= 25]: Distance step in meters.

maximum_dispersal_distance [= 3000]: Maximum dispersal distance in meters.

min_percent [= 1]: A minimum percent of seed bank to retain entry in seedBank element of forest.

stochastic_resampling [= FALSE]: A flag to indicate that stochastic resampling of stands is performed.

See also

Author

Miquel De Cáceres Ainsa, CREAF

Examples

default_watershed_control()
#> $watershed_model
#> [1] "tetis"
#> 
#> $weather_aggregation_factor
#> [1] 1
#> 
#> $tetis_parameters
#> $tetis_parameters$R_localflow
#> [1] 1
#> 
#> $tetis_parameters$R_interflow
#> [1] 1
#> 
#> $tetis_parameters$R_baseflow
#> [1] 10
#> 
#> 
#> $serghei_parameters
#> $serghei_parameters$input_dir
#> [1] ""
#> 
#> $serghei_parameters$output_dir
#> [1] ""
#> 
#> $serghei_parameters$force_equal_layer_widths
#> [1] FALSE
#> 
#> 
#> $dispersal_parameters
#> $dispersal_parameters$distance_step
#> [1] 25
#> 
#> $dispersal_parameters$maximum_dispersal_distance
#> [1] 3000
#> 
#> $dispersal_parameters$min_percent
#> [1] 1
#> 
#> $dispersal_parameters$stochastic_resampling
#> [1] FALSE
#> 
#>