Simulations for spatially-distributed forest stands

Functions that allow calling local models spwb, growth or fordyn, for a set of forest stands distributed in specific locations. No spatial processes are simulated.

spwbpoints(y, SpParams, meteo, localControl = defaultControl(),
       dates = NULL, keepResults = TRUE, summaryFunction = NULL, summaryArgs=NULL, 
       parallelize = FALSE, numCores = detectCores()-1, chunk.size = NULL, 
       progress = TRUE)
spwbpixels(y, SpParams, meteo, localControl = defaultControl(),
       dates = NULL, keepResults = TRUE, summaryFunction = NULL, summaryArgs=NULL, 
       parallelize = FALSE, numCores = detectCores()-1, chunk.size = NULL, 
       progress = TRUE)
spwbgrid(y, SpParams, meteo, localControl = defaultControl(),
       dates = NULL, keepResults = TRUE, summaryFunction = NULL, summaryArgs=NULL, 
       parallelize = FALSE, numCores = detectCores()-1, chunk.size = NULL, 
       progress = TRUE)
growthpoints(y, SpParams, meteo, localControl = defaultControl(),
       dates = NULL, keepResults = TRUE, summaryFunction = NULL, summaryArgs=NULL, 
       parallelize = FALSE, numCores = detectCores()-1, chunk.size = NULL, 
       progress = TRUE)
growthpixels(y, SpParams, meteo, localControl = defaultControl(),
       dates = NULL, keepResults = TRUE, summaryFunction = NULL, summaryArgs=NULL, 
       parallelize = FALSE, numCores = detectCores()-1, chunk.size = NULL, 
       progress = TRUE)
growthgrid(y, SpParams, meteo, localControl = defaultControl(),
       dates = NULL, keepResults = TRUE, summaryFunction = NULL, summaryArgs=NULL, 
       parallelize = FALSE, numCores = detectCores()-1, chunk.size = NULL, 
       progress = TRUE)
fordynpoints(y, SpParams, meteo, localControl = defaultControl(),
       dates = NULL, managementFunction = NULL, managementArgs = NULL,
       keepResults = TRUE, summaryFunction = NULL, summaryArgs=NULL, 
       parallelize = FALSE, numCores = detectCores()-1, chunk.size = NULL, 
       progress = TRUE)
fordynpixels(y, SpParams, meteo, localControl = defaultControl(),
       dates = NULL, managementFunction = NULL, managementArgs = NULL,
       keepResults = TRUE, summaryFunction = NULL, summaryArgs=NULL, 
       parallelize = FALSE, numCores = detectCores()-1, chunk.size = NULL, 
       progress = TRUE)
fordyngrid(y, SpParams, meteo, localControl = defaultControl(),
       dates = NULL, managementFunction = NULL, managementArgs = NULL,
       keepResults = TRUE, summaryFunction = NULL, summaryArgs=NULL, 
       parallelize = FALSE, numCores = detectCores()-1, chunk.size = NULL, 
       progress = TRUE)

Arguments

y

An object of class SpatialPointsLandscape-class, SpatialPixelsLandscape-class or SpatialGridLandscape-class, depending on the function called.

SpParams

A data frame with species parameters (see SpParamsMED).

meteo

Meteorology data (see details).

localControl

A list of local model control parameters (see defaultControl).

dates

A Date object with the days of the period to be simulated. If NULL, then the whole period of meteo is used.

keepResults

Boolean flag to indicate that point/cell simulation results are to be returned (set to FALSE and use summary functions for large data sets).

summaryFunction

An appropriate function to calculate summaries (e.g., summary.spwb).

summaryArgs

List with additional arguments for the summary function.

parallelize

Boolean flag to try parallelization (will use all clusters minus one).

numCores

Integer with the number of cores to be used for parallel computation.

chunk.size

Integer indicating the size of chuncks to be sent to different processes (by default, the number of spatial elements divided by the number of cores).

progress

Boolean flag to display progress information for simulations.

managementFunction

A function that implements forest management actions (see fordyn).

managementArgs

A list of arguments to be passed to the managementFunction (see fordyn).

Details

Simulation functions accept different formats for meteorological input (parameter meteo). The user may supply four kinds of weather sources:

An object of SpatialPointsMeteorology-class, SpatialGridMeteorology-class or SpatialPixelsMeteorology-class, with the same spatial geometry as y.

An object of MeteorologyInterpolationData-class.

A data frame with meteorological data common for all points or grid cells.

A character vector with filenames from which to read weather data.

In the case of (1) or (4), all the spatio-temporal variation of weather is already supplied by the user. In the case of (2), interpolation of weather is done for each spatial location and the period indicated in meteo or, if supplied, dates. Finally, in the case of (3) spatial variation of weather is not considered.

Value

A list of class of the same name as the function called, also inheriting from an appropriate summary class (summarypoints, summarypixels or summarygrid), containing four elements:

sp: An object with spatial information (of SpatialPoints-class, SpatialPixels-class or SpatialGrid-class).

xlist: A list of spwbInput or growthInput objects for each simulated stand, to be used in subsequent simulations (see updateState).

forestlist: A list of forest objects for each simulated stand (only in fordynpoints, fordynpixels and fordyngrid), to be used in subsequent simulations (see updateState).

resultlist: A list of model output for each simulated stand (if keepResults = TRUE).

summarylist: A list of model output summaries for each simulated stand (if summaryFunction is not NULL).

Author

Miquel De Cáceres Ainsa, CREAF

See also

Examples

if (FALSE) {
#Load example watershed (inherits from SpatialPixelsLandscape)
data("examplepointslandscape")

#Load example meteo data frame from package meteoland
data("examplemeteo")

#Load default medfate parameters
data("SpParamsMED")

#Perform simulation
dates = seq(as.Date("2001-03-01"), as.Date("2001-03-15"), by="day")
res = spwbpoints(examplepointslandscape, SpParamsMED, examplemeteo, dates = dates)

#Generate summaries (these could have also been specified when calling 'spwbpoints')
res_sum = summary(res, summaryFunction = summary.spwb, freq="month")

#Plot summaries
plot(res_sum, "Transpiration", "2001-03-01")
}