Displays spatial simulation summaries

Produces graphical output of the summaries of a simulation models

Usage

plot_summary(x, variable, date, r = NULL, ...)

Arguments

x

An object of class sf, with simulation summaries. Alternatively an object of class spwb_land, growth_land or fordyn_land.

variable

The variable to be drawn.

date

The date of the summary to be plotted.

r

An object of class SpatRaster, defining the raster topology.

...

Additional parameters (passed to scale definition, such as limits).

Value

An object of class ggplot.

Details

Appropriate input objects for x can be originated from calls to simulation_summary. Alternatively, if summary functions were specified at the time of performing simulations, the result of the spatial simulation function (e.g. spwb_spatial) will already contain the summaries. A special case is made for spwb_land, growth_land, and fordyn_land, that are accepted inputs as x, because their child element 'sf' is used.

See also

Author

Miquel De Cáceres Ainsa, CREAF.