Summarizes spatial simulation results

Creates spatial objects containing summaries of simulations

simulation_summary(object, summary_function, ...)

Arguments

object

An object of class 'sf' simulation results (e.g. the result of calling spwb_spatial).

summary_function

The summary function to be executed on simulation results (see details).

...

Additional parameters to the summary function.

Value

An object of class sf, with the following two elements:

geometry: Spatial geometry.

id: Stand id, taken from the input.

summary: A list of model output summaries for each simulated location.

Details

The function supplied should take as input an object of local simulation function, i.e. spwb, growth, or fordyn. The output should be a matrix with dates as rows and variables in columns. An example of suitable function is summary.spwb.

See also

Author

Miquel De Cáceres Ainsa, CREAF.