Watershed simulations

Function spwb_land implements a distributed hydrological model that simulates daily local water balance, from spwb_day, on grid cells of a watershed while accounting for overland runoff, subsurface flow and groundwater flow between cells. Function growth_land is similar, but includes daily local carbon balance and growth processes in grid cells, provided by growth_day.

spwb_land(
  y,
  SpParams,
  meteo,
  dates = NULL,
  summary_frequency = "years",
  local_control = medfate::defaultControl(),
  correction_factors = default_watershed_correction_factors(),
  progress = TRUE
)

growth_land(
  y,
  SpParams,
  meteo,
  dates = NULL,
  summary_frequency = "years",
  local_control = medfate::defaultControl(),
  correction_factors = default_watershed_correction_factors(),
  progress = TRUE
)

Arguments

y

An object of class sf containing watershed information.

SpParams

A data frame with species parameters (see SpParamsMED).

meteo

Input meteorological data (see section details).

dates

A Date object describing the days of the period to be modeled.

summary_frequency

Frequency in which summary layers will be produced (e.g. "years", "months", ...) (see cut.Date).

local_control

A list of control parameters (see defaultControl) for function spwb_day or growth_day.

correction_factors

A list of watershed correction factors for hydraulic parameters.

progress

Boolean flag to display progress information for simulations.

Value

Function spwb_land list of class 'spwb_land' with the following elements:

sf: An object of class sf, similar to the output of spwb_spatial, with the following columns:

state: A list of model input objects for each simulated stand.

id: Stand id, taken from the input.

aquifer: A numeric vector with the water volume in the aquifer of each cell.

snowpack: A numeric vector with the snowpack water equivalent volume of each cell.

summary: A list of cell summaries, containing the following variables:

Rain: Rainfall (in mm).

Snow: Snowfall (in mm).

Snowmelt: Snow melt (in mm).

Interception: Rainfall interception (in mm).

NetRain: Net rain, i.e. throughfall, (in mm).

Runon: The amount of water imported from other cells via surface runoff (in mm).

Runoff: The amount of water exported via surface runoff (in mm).

Infiltration: The amount of water infiltrating into the soil (in mm).

DeepDrainage: The amount of water draining from soil to the aquifer via deep drainage (in mm).

SaturationExcess: The amount of water that reaches the soil surface because of soil saturation (in mm).

AquiferDischarge: The amount of water that reaches deepest soil layer from a saturated aquifer (in mm).

SubsurfaceInput: The amount of water that reaches the soil from adjacent cells via subsurface flow (in mm).

SubsurfaceOutput: The amount of water that leaves the soil towards adjacent cells via subsurface flow (in mm).

GroundwaterInput: The amount of water that reaches the aquifer from adjacent cells via groundwater flow (in mm).

GroundwaterOutput: The amount of water that leaves the aquifer towards adjacent cells via groundwater flow (in mm).

SoilEvaporation: Bare soil evaporation (in mm).

Transpiration: Plant transpiration (in mm).

SWE: Snow water equivalent (in mm) of the snowpack.

Psi1: Soil water potential of the topmost layer (in MPa).

SoilVol: Soil water volume integrated across vertical layers (in mm).

WTD: Water table depth (in mm from surface).

watershed_balance: A data frame with as many rows as summary points and where columns are components of the water balance at the watershed level (i.e., rain, snow, interception, infiltration, soil evaporation, plant transpiration, ...).

watershed_soil_balance: A data frame with as many rows as summary points and where columns are components of the water balance at the watershed level restricted to those cells with a soil definition.

daily_runoff: A matrix with daily runoff (in m3/day) at each of the outlet cells of the landscape.

Details

Functions spwb_land and growth_land require daily meteorological data over grid cells. The user may supply four different inputs:

A data frame with meteorological data common for all spatial location (spatial variation of weather not considered).

An object of class stars with interpolation data, created by package meteoland.

DEPRECATED: An object of SpatialPixelsMeteorology-class or SpatialGridMeteorology-class. All the spatio-temporal variation of weather is already supplied by the user.

DEPRECATED: An object of MeteorologyInterpolationData-class. Interpolation of weather is performed over each spatial unit every simulated day.

See also

Author

Miquel De Cáceres Ainsa, CREAF.

Examples

if (FALSE) {
# Load example watershed data
data("examplewatershed")

# Transform example to 'sf' 
y = sp_to_sf(examplewatershed)
  
# Load example meteo data frame from package meteoland
data("examplemeteo")
  
# Load default medfate parameters
data("SpParamsMED")
  
# Set simulation period
dates = seq(as.Date("2001-01-01"), as.Date("2001-03-31"), by="day")

# Launch simulations
res = spwb_land(y, SpParamsMED, examplemeteo, dates = dates, summary_frequency = "month")
}